Revert "Bluerov pitchprediction"
