--- meshes/bluerov_white.mtl
@@ -1,2 +1,0 @@
-# Blender 3.4.1 MTL File: 'bluerov.blend'
-# www.blender.org
