+++ src/full-sim/ablation_study/plot_ablation_results_manual_input.m
@@ -1,23 +1,16 @@
-rmses = [.0414, .0411, .0416; ...
-    .0407, .0408, .0407; ...
-    .0409, .0406, .0403; ...
-    .0401, .0412, .0407; ...
-    .0423, .0426, .0438; ...
-    .0447, .0443, .0440; ...
-    .0529, .0545, .0521];
-rmse_avgs = mean(rmses, 2)
-rmse_stdevs = std(rmses, 0, 2)
-num_inputs = [24, 20, 17, 14, 10, 6, 3];
+new_rmses = [.0203, .0177, .0163, .0151, .0144, .0142, .0135, .0136, .0140];
+new_rmse_stdevs = [0, 0., 0, 0, 0, 0, 0, 0, 0];
+num_inputs = [40, 37, 34, 26, 22, 14, 12, 8, 6];
     
 figure
 % scatter(num_inputs, loss_avgs, 'MarkerEdgeColor', '#56B4E9', 'LineWidth', 2)
 hold on
-er_bars = errorbar(num_inputs, rmse_avgs, rmse_stdevs); %'Color', '#56B4E9')
+er_bars = errorbar(num_inputs, new_rmses, new_rmse_stdevs); %'Color', '#56B4E9')
 my_xlab = xlabel("Number of Channel Inputs");
 my_ylab = ylabel("Prediction RMSE");
 set(gca, 'XDir', 'reverse')
-ylim([.039, .055])
-xlim([0, 25])
+ylim([.013, .0195])
+xlim([0, 42])
 % my_title = title('Network Performance on Decreasing Input Channels');
 
 % set(fit_line, ...
@@ -49,11 +42,10 @@ set(gca, ...
   'XMinorTick'  , 'on'      , ...
   'YMinorTick'  , 'on'      , ...
   'YGrid'       , 'on'      , ...
-  'YTick'       , [.04:.002:.056], ...
   'XGrid'       , 'off'     , ...
   'XColor'      , [.3 .3 .3], ...
   'YColor'      , [.3 .3 .3], ...
   'LineWidth'   , 1         );
 
 set(gcf, ...
-    'Position', [100 100 600 275]);
+    'Position', [100 100 600 250]);
