+++ src/full-sim/data-generation/PIDCtlr.jl
@@ -1,70 +1,44 @@
-using RigidBodyDynamics, Distributions, Random
+module PIDCtlr
+export CtlrCache, pid_control!
+
+using RigidBodyDynamics
+
+include("TrajGenMain.jl")
+include("StructDefs.jl")
 
 # ------------------------------------------------------------------------
 #                              SETUP 
 # ------------------------------------------------------------------------
-arm_Kp =  3.38e-2 #.304e-2
-arm_Ki = 4.39e-1 #9.38e-2
-arm_Kd = 1.74e-3 #9.63e-4
-v_Kp = 1.2 #3.
-v_Ki = 0.6 #3.6
-v_Kd = 1.61 #1.68
-Kp = [2.0, v_Kp, v_Kp, v_Kp, arm_Kp, 4.59e-2, 1.35e-2, 5.4e-4] #, 20.]
-Ki = [1., v_Ki, v_Ki, v_Ki, arm_Ki, 3.73e-1, 9.64e-2, 3.6e-3] #3.037e-6] #, .1]
-Kd = [0.1, v_Kd, v_Kd, v_Kd, arm_Kd, 3.82e-3, 1.27e-3, 2.03e-5] #, .002]
-
-torque_lims = [20., 71.5, 88.2, 177., 10.0, 10.0, 10.0, 0.6] #, 600]
-
-# Sensor noise distributions 
-# Implemented:
-# Encoder --> joint position noise -integration-> joint velocity noise
-# Gyroscope --> vehicle body vel noise 
-v_ang_vel_noise_dist = Distributions.Normal(0, .0013) # 75 mdps (LSM6DSOX)
-arm_pos_noise_dist = Distributions.Normal(0, .0017/6) # .1 degrees, from Reach website
-accel_noise_dist = Distributions.Normal(0, 0.017658/10) # 1.8 mg = .0176 m/s2 (LSM6DSOX)
-
-gyro_rand_walk_dist = Distributions.Normal(0, .000001)
-accel_rand_walk_dist = Distributions.Normal(0, 0)#0.00001)
+arm_Kp = 30.
+arm_Kd = 0.05
+arm_Ki = 9. 
+v_Kp = 50.
+v_Kd = 1.0
+v_Ki = 1.25
+default_Kp = [v_Kp, v_Kp, v_Kp, v_Kp, arm_Kp, arm_Kp, arm_Kp, 20.]
+default_Kd = [v_Kd, v_Kd, v_Kd, v_Kd, arm_Kd, arm_Kd, arm_Kd, 0.002]
+default_Ki = [v_Ki, v_Ki, v_Ki, v_Ki, arm_Ki, arm_Ki, arm_Ki, 0.1]
+default_torque_lims = [20., 71.5, 88.2, 177., 10.0, 10.0, 10.0, 0.6]
 
 mutable struct CtlrCache
+    Kp::Array{Float64}
+    Kd::Array{Float64}
+    Ki::Array{Float64}
     time_step::Float64
-    ctrl_freq::Float64
-    ctrl_steps::Float64
     vel_error_cache::Array{Float64}
     vel_int_error_cache::Array{Float64}
     step_ctr::Int
     joint_vec
     des_vel::Array{Float64}
+    tau_lims::Array{Float64}
     taus
-    noisy_qs
-    noisy_vs
-    filtered_vs
-    last_v̇
-    filtered_state
-    rand_walks
-    swap_times
-    traj_num
     
-    function CtlrCache(dt, control_frequency, state, swap_times)
-        mechanism = state.mechanism
-        if length(joints(mechanism)) == 5
-            vehicle_joint, jointE, jointD, jointC, jointB = joints(mechanism)
-            joint_vec = [vehicle_joint, jointE, jointD, jointC, jointB]
-            num_actuated_dofs = 8
-            num_dofs = 10
-        elseif length(joints(mechanism)) == 6
-            vehicle_joint, jointE, jointD, jointC, jointB, jawjoint = joints(mechanism)
-            joint_vec = [vehicle_joint, jointE, jointD, jointC, jointB, jawjoint]
-            num_actuated_dofs = 9
-            num_dofs = 11
-        end
-        ctrl_loop_num_steps = 4*(1/dt)/control_frequency
-        new(dt, control_frequency, ctrl_loop_num_steps, #=
-        =# zeros(num_actuated_dofs), zeros(num_actuated_dofs), 0, joint_vec, #=
-        =# zeros(num_actuated_dofs), Array{Float64}(undef, num_dofs, 1), #=
-        =# configuration(state), velocity(state), #=
-        =# zeros(num_dofs), zeros(num_dofs), 
-        MechanismState(mechanism), zeros(6), swap_times, 1) 
+    function CtlrCache(dt, mechanism)
+        vehicle_joint, jointE, jointD, jointC, jointB = joints(mechanism)
+        joint_vec = [vehicle_joint, jointE, jointD, jointC, jointB]
+        new(default_Kp, default_Kd, default_Ki, #=
+        =# dt, zeros(8), zeros(8), 0, joint_vec, #=
+        =# zeros(8), default_torque_lims, Array{Float64}(undef, 10, 1)) 
     end
 end
 
@@ -76,25 +50,16 @@ Given a `torque`` value and a torque `limit``, bounds `torque`` to be within the
 
 Assumes the torque limit is the same in the positive and negative directions.
 """
-function impose_torque_limit(torque, limit)
+function impose_torque_limit!(torque, limit)
     if torque > limit
-        new_tau = limit
+        torque = limit
     elseif torque < -limit
-        new_tau = -limit
-    else
-        new_tau = torque
+        torque = -limit
     end
-    new_tau
 end
 
 function resettimestep(c::CtlrCache)
     c.step_ctr = 0
-end
-
-function skew(x1, x2, x3)
-    output = [0 -x3 x2;
-            x3 0 -x1;
-            -x2 x1 0]
 end
 
 # ------------------------------------------------------------------------
@@ -107,112 +72,48 @@ Requires the parameters of the trajectory to be followed (pars=trajParams), whic
 Only happens every 4 steps because integration is done with Runge-Kutta.
 """
 # function pid_control!(torques::AbstractVector, t, state::MechanismState, pars, c)
-function pid_control!(torques::AbstractVector, t, state::MechanismState, pars, c, result, h_wrenches)
+function pid_control!(torques::AbstractVector, t, state::MechanismState, pars, c)
     # If it's the first time called in the Runge-Kutta, update the control torque
     # println("Made it inside the function! Ctr = $(time_step_ctr)")
     if rem(c.step_ctr, 4) == 0
-        # println("Joint E Torques: $(torques[7])")
         # Set up empty vector for control torques
-        c_taus = zeros(size(c.taus, 1),1)
+        c_taus = zeros(10,1)
         if c.step_ctr == 0
-            torques[6] = 5.2 # ff z value
+            torques[6] = 7. # 8
             torques[3] = 0.
             torques[5] = 0.
-            torques[4] = -2.3 # ff x value
-            torques[7] = -.002 # ff joint E value
-            torques[8] = -.32255 # ff Joint D value 
-            torques[9] = -.0335 # ff Joint C value
-            torques[10] = 0 #.5e-5
-            println("At time... ")
+            torques[4] = -1. # -1.8
         end
         
-        # Roll and pitch are not controlled
+        # Roll and pitch are not controlled.
         torques[1] = 0.
         torques[2] = 0.
         
         # println("Requesting des vel from trajgen")
         # c.des_vel = TrajGen.get_desv_at_t(t, pars)
+        c.des_vel = TrajGen.get_desv_at_t(t, pars)
+        # Don't move the manipulator
+        # c.des_vel[end-3:end] = zeros(4,1)
+        # println("Got desired velocity")
 
-        # if rem(c.step_ctr, 1000) == 0
-        #     @show c.rand_walks
-        #     # @show result.jointwrenches
-        # #     # println("Desired velocity vector: $(c.des_vel)")
-        # #     wrist_wrenches = result.totalwrenches[BodyID(wrist_body)]
-        # #     wrist_wrench_wrist_frame = transform(wrist_wrenches, relative_transform(state, base_frame, wrist_frame))
-        # #     @show wrist_wrench_wrist_frame
-        # end
-
-        if rem(c.step_ctr, c.ctrl_steps) == 0 && c.step_ctr != 0
-
-            if t > c.swap_times[c.traj_num]
-                c.traj_num += 1
-            end
-            mod_time = c.traj_num == 1 ? t : t-c.swap_times[c.traj_num-1]
-            c.des_vel = get_desv_at_t(mod_time, pars[c.traj_num])
-    
-            # ff_torques = dynamics_bias(state, h_wrenches)
-
-            noisy_poses = similar(configuration(state))
-            noisy_vels = similar(velocity(state))
-            fill!(noisy_poses, 0.)
-            fill!(noisy_vels, 0.)
-
-            c.rand_walks[1:3] .+= rand(gyro_rand_walk_dist, 3)
-            c.rand_walks[4:6] .+= rand(accel_rand_walk_dist, 3)
-
-            last_noisy_R = Rotations.QuatRotation(c.noisy_qs[1:4,end])
-            # println("Starting to add noise... Starting noisy poses")
-            # @show velocity(state)
-            # @show noisy_vels
-            add_arm_noise!(noisy_poses, noisy_vels, configuration(state)[8:end], c.noisy_qs[8:end,end], 1/c.ctrl_freq)
-            # @show noisy_vels
-            new_noisy_R = add_rotational_noise!(noisy_poses, noisy_vels, velocity(state)[1:3], last_noisy_R, 1/c.ctrl_freq, c.rand_walks)
-            # @show noisy_vels
-            add_linear_noise!(noisy_poses, noisy_vels, result.v̇[1:6], c.last_v̇, c.noisy_vs[1:6,end], c.noisy_qs[5:7,end], last_noisy_R, new_noisy_R, 1/c.ctrl_freq, c.rand_walks)
-            # @show noisy_vels
-            # if rem(c.step_ctr, 1000) == 0
-            #     @show result.v̇[6]
-            #     @show result.accelerations[BodyID(3)]
-            # end
-
-            c.last_v̇ = result.v̇
-            c.noisy_vs = cat(c.noisy_vs, noisy_vels, dims=2)
-            c.noisy_qs = cat(c.noisy_qs, noisy_poses, dims=2)
-
-            filtered_velocity = moving_average_filter_velocity(5, c.noisy_vs, velocity(state))
-            c.filtered_vs = cat(c.filtered_vs, filtered_velocity, dims=2)
-            filtered_vels = similar(velocity(state))
-            filtered_vels[:] = filtered_velocity
-
-            set_configuration!(c.filtered_state, noisy_poses)
-            set_velocity!(c.filtered_state, filtered_vels)
-            ff_torques = dynamics_bias(c.filtered_state, h_wrenches)
-
-            # Get forces for vehicle (yaw, surge, sway, heave)
-            for dir_idx = 3:6
-                # println("PID ctlr on vehicle")
-                actual_vel = velocity(state, c.joint_vec[1])
-                ctlr_tau = PID_ctlr(torques[dir_idx][1], t, filtered_velocity[dir_idx], dir_idx, c, ff_torques)
-                # ctlr_tau = torques[dir_idx]
-                c_taus[dir_idx] = ctlr_tau 
-                torques[dir_idx] = ctlr_tau
-            end
-            
-            # Get torques for the arm joints
-            for jt_idx in 2:length(c.joint_vec) # Joint index (1:vehicle, 2:baseJoint, etc)
-                idx = jt_idx+5 # velocity index (7 to 10)
-                ctlr_tau = PID_ctlr(torques[idx][1], t, filtered_velocity[idx], idx, c, ff_torques) 
-                torques[velocity_range(state, c.joint_vec[jt_idx])] .= [ctlr_tau] 
-                c_taus[idx] = ctlr_tau 
-            end
-            #TODO switch to push! ?
-            c.taus = cat(c.taus, c_taus, dims=2)
-            # push!(c.taus, copy(c_taus))
+        # Get forces for vehicle (yaw, surge, sway, heave)
+        for dir_idx = 3:6
+            # println("PID ctlr on vehicle")
+            ctlr_tau = PID_ctlr(torques[dir_idx][1], t, velocity(state, c.joint_vec[1])[dir_idx], dir_idx, c)
+            c_taus[dir_idx] = ctlr_tau 
+            torques[dir_idx] = ctlr_tau
         end
-    end
-    if rem(c.step_ctr, 4000) == 0
-        # println("At time $(c.step_ctr/4000)...")
-        print("$(c.step_ctr/4000)... ")
+        
+        # Get torques for the arm joints
+        for jt_idx in 2:5 # Joint index (1:vehicle, 2:baseJoint, etc)
+            # println("PID ctlr on arm")
+            idx = jt_idx+5 # velocity index 
+            ctlr_tau = PID_ctlr(torques[idx][1], t, velocity(state, c.joint_vec[jt_idx]), idx, c) 
+            c_taus[jt_idx+5] = ctlr_tau
+            damp_tau = -0.1*velocity(state, c.joint_vec[jt_idx])
+            torques[velocity_range(state, c.joint_vec[jt_idx])] .= [ctlr_tau] + damp_tau
+        end
+        c.taus = cat(c.taus, c_taus, dims=2)
     end
     c.step_ctr = c.step_ctr + 1
 end;
@@ -228,52 +129,30 @@ Imposes a PID controller on one joint.
 
 Returns a controller torque value, bounded by the torque limits and some dτ/dt value. 
 """
-function PID_ctlr(torque, t, vel_act, idx, c, ff)
-    dt = 1/c.ctrl_freq 
+function PID_ctlr(torque, t, vel_act, idx, c)
     actuated_idx = idx-2
     d_vel = c.des_vel[actuated_idx]
     vel_error = vel_act[1] - d_vel
-    d_vel_error = (vel_error - c.vel_error_cache[actuated_idx])/dt
+    d_vel_error = (vel_error - c.vel_error_cache[actuated_idx])/c.time_step
     # println("D_Velocity error on idx$(j_idx): $(d_vel_error)")
-    c.vel_int_error_cache[actuated_idx] += vel_error*dt
-
-    p_term = -Kp[actuated_idx]*vel_error
-    d_term = - Kd[actuated_idx]*d_vel_error
-    i_term = - Ki[actuated_idx]*c.vel_int_error_cache[actuated_idx]
-    d_tau = p_term + d_term + i_term
+    c.vel_int_error_cache[actuated_idx] = c.vel_int_error_cache[actuated_idx] + vel_error*c.time_step
+    d_tau = -c.Kp[actuated_idx]*vel_error - c.Kd[actuated_idx]*d_vel_error - c.Ki[actuated_idx]*c.vel_int_error_cache[actuated_idx]
     # println("Ideal tau: $(d_tau)")
 
-    # Can only change torque a small amount per time step 
-    # arm joints can change faster than thrusters
-    #TODO if ever change controller or actual time step, change limits
-    if actuated_idx < 5 # vehicle joints
-        lim = .001 # N per 0.01s
-    elseif actuated_idx < 8 # arm joints
-        lim = .1 
-    else # wrist joint
-        lim = 0.006
-    end
-    tau_diff_prev_to_inv_dyn = .25ff[idx] - .25torque
-    d_tau_w_ff = limit_d_tau(tau_diff_prev_to_inv_dyn+d_tau, lim)
-
-    # Torque limits
-    if actuated_idx >= 5
-        new_tau = torque .+ d_tau_w_ff
-        # new_tau = torque .+ d_tau
+    # Can only change torque a small amount per time step
+    if 5 <= actuated_idx <= 8
+        # if it's an arm joint
+        lim = 0.01
     else
-        d_tau = limit_d_tau(d_tau, lim)
-        new_tau = torque .+ d_tau
-    end 
+        # if it's a vehicle "thruster"
+        lim = 0.001
+    end
+    d_tau = limit_d_tau(d_tau, lim)
     
-    new_tau = impose_torque_limit(new_tau, torque_lims[actuated_idx])
+    # Torque limits
+    new_tau = torque .+ d_tau
+    impose_torque_limit!(new_tau, c.tau_lims[actuated_idx])
 
-    # if rem(c.step_ctr, 100) == 0 && actuated_idx == 7
-    #     @show new_tau
-    # end
-    # if  t > 12 && rem(c.step_ctr, 100) == 0
-    #     @show actuated_idx
-    #     @show new_tau
-    # end
     # # store velocity error term
     c.vel_error_cache[actuated_idx]=vel_error
     # c.vel_int_error = c.vel_int_error + vel_error
@@ -290,38 +169,4 @@ function limit_d_tau(d_tau, limit)
     return d_tau
 end
 
-function add_arm_noise!(noisy_poses, noisy_vels, joint_poses, last_noisy_joint_pose, dt)
-    noisy_joint_poses = joint_poses + rand(arm_pos_noise_dist, length(joint_poses)) 
-    # noisy_joint_poses = joint_poses
-    # noisy_joint_poses[2:4] = joint_poses[2:4] + rand(arm_pos_noise_dist, 3)
-    noisy_velocity = (noisy_joint_poses - last_noisy_joint_pose)./dt
-    noisy_poses[8:end] .= noisy_joint_poses
-    noisy_vels[7:end] .= noisy_velocity
-end
-
-function add_rotational_noise!(noisy_poses, noisy_vels, body_vels, last_R, dt, rand_walk)
-    noisy_vels[1:3] .= body_vels[1:3] + rand(v_ang_vel_noise_dist, 3) + rand_walk[1:3]
-    new_R = last_R*exp(skew(noisy_vels[1:3]...)*dt)
-    new_noisy_R = Rotations.RotMatrix(SMatrix{3,3}(new_R))
-    quat_new_noisy_R = Rotations.QuatRotation(new_noisy_R)
-    noisy_poses[1:4] .= [quat_new_noisy_R.w, quat_new_noisy_R.x, quat_new_noisy_R.y, quat_new_noisy_R.z]
-    return quat_new_noisy_R
-end
-
-function add_linear_noise!(noisy_poses, noisy_vels, v̇, last_v̇, last_body_vels, last_pos, last_R, new_R, dt, rand_walk)
-    noisy_linear_accels = last_v̇[4:6] + rand(accel_noise_dist, 3) + rand_walk[4:6]
-    noisy_vels[4:6] .= last_body_vels[4:6] + noisy_linear_accels.*dt #+last_v̇[4:6])./2 .*dt
-    noisy_vels_in_space = last_R*SVector{3}(last_body_vels[4:6]) #+ new_R*SVector{3}(noisy_vels[4:6]))./2
-    noisy_poses[5:7] .= last_pos + noisy_vels_in_space.*dt   
-end
-
-function moving_average_filter_velocity(filt_size, noisy_vs, act_vs)
-    num_its = size(noisy_vs, 2)
-    if num_its < filt_size
-        filtered_vels = act_vs
-    else
-        filtered_vels = sum(noisy_vs[:,end-filt_size+1:end], dims=2) ./ filt_size  
-    end
-    # @show filtered_vels
-    return filtered_vels
 end
