+++ src/full-sim/data-generation/SimWExt.jl
@@ -16,23 +16,14 @@ function simulate_with_ext_forces(state0::MechanismState{X}, final_time, pars, c
             # println("Current State:")
             # println(configuration(state))
             hydro_calc!(hydro_wrenches, t, state)
+            control!(control_torques, t, state, pars, ctlr)
+            # println("Control torques")
+            # println(control_torques)
+
+            # hydro_calc!(hydro_wrenches, t, state)
             # println("Hydro wrenches")
             # println(hydro_wrenches)
 
-            control!(control_torques, t, state, pars, ctlr, result, hydro_wrenches)
-            # println("Control torques")
-            # println(control_torques)
-            # println("--------------- NEW ITERATION -------------------")
-            # println("prev result")
-            # println(result)
-
-            # println("State")
-            # println(configuration(state))
-            # println(velocity(state))
-            # println("Control torques:")
-            # println(control_torques)
-            # println("Hydro Wrenches")
-            # println(hydro_wrenches)
             dynamics!(result, state, control_torques, hydro_wrenches; stabilization_gains=stabilization_gains)
             # println("result:")
             # println(result.v̇)
