+++ src/full-sim/data-generation/StructDefs.jl
@@ -1,0 +1,9 @@
+mutable struct jointState
+    θs::Array{Float64}
+    dθs::Array{Float64}
+end
+
+mutable struct Waypoints
+    start::jointState 
+    goal::jointState
+end
