+++ src/full-sim/helper-funcs/get_remaining_idxs.m
@@ -9,8 +9,12 @@ Returns:
 rem_idxs: list of indices of remaining feature groups. 
 (1:41, minus the channels for the eliminated groups.)
 %}
-function rem_idxs = get_remaining_idxs(elimd_gps, chan_idxs)   
-    all_idxs = 21:1:44;
+function rem_idxs = get_remaining_idxs(elimd_gps)
+    load('data/channel_dict.mat', 'chan_idxs')
+    chan_idxs = rmfield(chan_idxs, 'pitch');
+    chan_idxs = rmfield(chan_idxs, 'dt');
+    
+    all_idxs = 1:1:41;
 
     for i = 1:length(elimd_gps)
         group_name = elimd_gps(i);
