+++ src/full-sim/helper-funcs/import_and_parsing/ID_outliers.m
@@ -1,7 +1,6 @@
 function bad_idxs = ID_outliers(sequence_data)
 
     max_vals = [];
-%     disp(numel(sequence_data))
     for n = 1:numel(sequence_data)
         max_vals(n) = max(sequence_data{n}, [], 'all');
     end
@@ -10,7 +9,7 @@ function bad_idxs = ID_outliers(sequence_data)
 
     bad_idxs = [];
     for idx = 1:length(max_vals)
-        if max_vals(idx) > 6.0
+        if max_vals(idx) > 5.0
             bad_idxs = [bad_idxs idx];
         end
     end
