+++ src/full-sim/k_length_study/network_generation/k_length_study_redo.m
@@ -1,24 +1,18 @@
 %% Setup
 % Load data set
-load("data/full-sim-data-022223/FullData_50Hz.mat")
+load("data/full-sim-data-110822/FullData.mat")
 
 % get indices of remaining feature groups
-load('data/full-sim-data-022223/channel_dict.mat')
-pitch_idx = chan_idxs.act_pitch;
-chan_idxs = rmfield(chan_idxs, {'act_rpy', 'act_xyz', 'act_joint_pos', 'act_angular_vels', 'act_linear_vels', 'act_joint_vels', 'act_pitch'});
-
-% Make a list of all channel indices
-all_idxs = 21:1:44;
-elimd_gps = ["meas_linear_vels", "meas_joint_vels", "meas_xyz"];
-all_idxs = get_remaining_idxs(elimd_gps, chan_idxs);
+elimd_gps = ["xyz_poses", "xyz_vels", "goal_poses", "manip_des_vels", "goal_vels"];
+all_idxs = get_remaining_idxs(elimd_gps);
 
 %%
 % Initialize constants
 % ks =[5, 10, 15, 20, 25, 30, 40, 50, 60, 70, 75, 80, 90, 100, 125, 150, 175, 200];
 k = 25;
+pitch_idx = 23;
 numUnits = 384;
-%stretches = [1, 2, 3, 4, 5, 6, 7, 8];
-stretches = [5, 7]
+stretches = [1, 2, 3, 4, 5, 6, 7, 8];
 
 all_losses = [];
 subgroup_losses = [];
@@ -55,13 +49,13 @@ for idx = 1:length(stretches)
         OutputNetwork='best-validation-loss');
 
     %% Train the net
-   for take_n = 1:2
+   for take_n = 1:3
         [net, info] = trainNetwork(Inputs_Train,Resp_Train,layers,init_options);
         
         subgroup_losses = [subgroup_losses, info.FinalValidationLoss];
         subgroup_RMSEs = [subgroup_RMSEs, info.FinalValidationRMSE];
         %     
-        outputFile = fullfile("data/networks/iros-nets/simple_w_stretch_factor", strcat('stretch_', string(sf), '_take_', string(take_n), '.mat'));
+        outputFile = fullfile("data/networks/icra-redo-nets/simple_w_stretch_factor", strcat('stretch_', string(sf), '_take_', string(take_n), '.mat'));
 %         outputFile = fullfile("data/networks/full-nets", strcat('pre-ablationtest_', string(take_n), '.mat'));
         save(outputFile, 'net', 'info');
    end
