+++ src/full-sim/retrain_net.m
@@ -1,31 +1,35 @@
 %% Import data
 
 % Load data set
-load('data/full-sim-data-022223/FullData_10Hz.mat')
+load('data/full-sim-data-110822/FullData_10Hz.mat')
 
 % Load network to retrain
-load('data/networks/iros-nets/SingleStepNet_10Hz.mat')
+load('data/networks/icra-redo-nets/SingleStepNet_10Hz.mat')
 
 % Load validation set definition
-load('data/full-sim-data-022223/val_set.mat')
+load('data/full-sim-data-110822/val_set_10Hz.mat')
 
 
 %%
-load('data/full-sim-data-022223/channel_dict.mat')
-pitch_idx = chan_idxs.act_pitch;
-chan_idxs = rmfield(chan_idxs, {'act_rpy', 'act_xyz', 'act_joint_pos', 'act_angular_vels', 'act_linear_vels', 'act_joint_vels', 'act_pitch'});
+load('data/channel_dict.mat')
+chan_idxs = rmfield(chan_idxs, 'pitch');
+chan_idxs = rmfield(chan_idxs, 'dt');
 
-% Make a list of all channel indices
-all_idxs = 21:1:44;
-elimd_gps = ["meas_linear_vels", "meas_xyz", "meas_joint_vels"];
-all_idxs = get_remaining_idxs(elimd_gps, chan_idxs);
-input_idxs = all_idxs;
+elimd_gps = ["xyz_poses", "xyz_vels", "goal_poses", "manip_des_vels", "goal_vels"];
+
+all_idxs = 1:1:41;
+% Take out 
+for i = 1:length(elimd_gps) 
+    group_name = elimd_gps(i);
+    all_idxs = all_idxs(~ismember(all_idxs, chan_idxs.(group_name)));
+    chan_idxs = rmfield(chan_idxs, group_name);
+end
+
 
 %% Initialization
 
 % k = 5;     % Number of time steps to forecast (0.5s)
-% ks = [5 10 20 30 40];
-ks = [15 25 35];
+ks = [5 10];
 mbatch = 16;
 num_trajs_before_update = 16;
 val_freq = 50;
@@ -33,6 +37,7 @@ train_to_val_ratio = val_freq*(num_trajs_before_update/mbatch);
 num_epochs = 50;
 save_freq = 5; %epoc5
 init_learning_rate = .0005;
+pitch_idx = 13;
 
 retrain_options = trainingOptions("adam", ...
     InitialLearnRate= init_learning_rate,...
@@ -43,21 +48,25 @@ retrain_options = trainingOptions("adam", ...
     GradientThreshold=0.9,...
     ExecutionEnvironment="auto");
 
-sorted_XTest_10hz = sort_data_by_length(XTest);
-val_XTest = sorted_XTest_10hz(val_idxs);
+XTest = XTest_10hz;
+[~,I] = sort(cellfun(@length,XTest));
+XTest = XTest(I);
+XTest = XTest(3:end);
 
-sorted_XTrain_10hz = sort_data_by_length(XTrain);
+for i = 1:numel(XTest)
+    XTest{i} = XTest{i}(all_idxs, :);
+end
 
-% for i = 1:numel(XTest)
-%     XTest{i} = XTest{i}(all_idxs, :);
-% end
+XTest = XTest(val_idxs);
 
 % disp(numel(XTrain_10hz))
-% for i = 1:numel(XTrain)
-%     XTrain{i} = XTrain{i}(all_idxs, :);
-% end
+for i = 1:numel(XTrain_10hz)
+    XTrain{i} = XTrain_10hz{i}(all_idxs, :);
+end
 
-% XTrain = XTrain(40:end);
+[~,I] = sort(cellfun(@length,XTrain));
+XTrain = XTrain(I);
+XTrain = XTrain(40:end);
 % disp(size(XTrain{40}, 2))
 % num_rec_vars = size(XTrain{1}, 1);
 
@@ -69,10 +78,10 @@ sorted_XTrain_10hz = sort_data_by_length(XTrain);
 
 %% Train on predictions
 for k_val = 1:length(ks)
-    load('data/networks/iros-nets/SingleStepNet_10Hz.mat')
+    load('data/networks/icra-redo-nets/SingleStepNet_10Hz.mat')
     k = ks(k_val);
 %     validate = @(net) validate_net(net, XTest_subset, val_ns, k, p);
-    validate = @(net) validate_pitch_on_forecast_only(net, val_XTest, val_ns_10hz, k, input_idxs, pitch_idx);
+    validate = @(net) validate_pitch_on_forecast_only(net, XTest, val_ns, k, p, pitch_idx);
     first_error = validate(net)
     error_vec = [first_error]; % SWITCH THIS BACK
     error_vec_its = [1];      % SWITCH THIS BACK
@@ -85,7 +94,7 @@ for k_val = 1:length(ks)
     
     for epoch_n = 1:num_epochs
     
-        for retrain_idx = 1:290
+        for retrain_idx = 1:296
         
             traj_indices = [];
             trajs = {};
@@ -97,7 +106,7 @@ for k_val = 1:length(ks)
             end
             
             parfor it_num = 1:num_trajs_before_update
-                data = trajs{it_num}(input_idxs,:);
+                data = trajs{it_num};
                 top_n_lim = size(data, 2) - k - 1;
                 if top_n_lim <= k + 1
                     n = top_n_lim;
@@ -106,11 +115,10 @@ for k_val = 1:length(ks)
                 end
                 
                 % Generate a prediction
-                pred = full_forecast_norecur(net, data, n, k);
+                pred = full_forecast_norecur(net, data, n, k, p);
         
                 preds{it_num} = pred;
                 g_truth{it_num} = data(:,2:n+k);
-                g_truth{it_num}(2,:) = data(pitch_idx,2:n+k)
             end
         
             nan_idxs = ID_nans(preds);
@@ -136,7 +144,7 @@ for k_val = 1:length(ks)
         end
     
         if rem(epoch_n, save_freq) == 0 || epoch_n > 45
-            outputFile = fullfile(strcat("data/networks/iros-nets/10Hz_k", string(k)), strcat("take1_", string(epoch_n), "epochs.mat"));
+            outputFile = fullfile(strcat("data/networks/icra-redo-nets/10Hz_k", string(k)), strcat("take1_", string(epoch_n), "epochs.mat"));
             end_it = total_it-1;
             these_epochs_training_RMSE_vec = training_rmse_vec(start_it:end_it);
             save(outputFile, 'net', 'info', "error_vec", "error_vec_its", "these_epochs_training_RMSE_vec", "start_it", "end_it");
@@ -159,18 +167,18 @@ end
 % outputFile = fullfile("data/networks/full-nets/retrained_nets", 'test_net.mat');
 % save(outputFile, 'net', 'info', 'error_vec', 'training_rmse_vec');
 
-% function error = validate_net(net, X_test, ns, k, p)
-%     error = 0;
-%     parfor i = 1:numel(X_test)
-%         pred = full_forecast_norecur(net, X_test{i}, ns(i), k);
-% %         size(pred)
-%         pred = pred(:,end-k+1:end);
-%         g_truth = X_test{i}(:,ns(i)+1:ns(i)+k);
-%         rmse = sqrt(immse(pred, single(g_truth)));
-%         error = error + rmse;
-%     end
-%     error = error/numel(X_test);
-% end
+function error = validate_net(net, X_test, ns, k, p)
+    error = 0;
+    parfor i = 1:numel(X_test)
+        pred = full_forecast_norecur(net, X_test{i}, ns(i), k, p);
+%         size(pred)
+        pred = pred(:,end-k+1:end);
+        g_truth = X_test{i}(:,ns(i)+1:ns(i)+k);
+        rmse = sqrt(immse(pred, single(g_truth)));
+        error = error + rmse;
+    end
+    error = error/numel(X_test);
+end
